リファクタ: 利用されていない機能を削除や移動
不要な部分を削除

利用されていない機能を別ファイルに移動

リファクタ

リファクタ
