+++ Editor/RoadNetwork/EditingSystem/RoadNetworkEditingTrafficSignal.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 11e98b8d31004c4fa147b47b64af0bd8
+timeCreated: 1735712841
